refactor(quil-node): extract sibling modules from main.rs (Batches 1-2 + partial 3)
Mechanical refactor of crates/quil-node/src/main.rs per
.claude/plans/deep-bouncing-cerf.md. No behavior changes.

Batch 1 — Foundations:
  - util/multiaddr.rs: pure host:port multiaddr helpers
  - blossomsub_consensus_publisher.rs: ConsensusPublisher impl
  - dht_node.rs: run_dht_node entry point
  - worker_node.rs: run_worker_node entry point (~370 lines moved verbatim)

Batch 2 — Slim main():
  - diagnostic.rs: --peer-id / --node-info / --peer-info / --metrics
    handlers + --import-db block. main() shrinks ~155 lines.
  (Step 5 — bulk move of run_master_node to master_node/mod.rs —
  deferred to the final batch; phases are extracted from main.rs
  in place, and the residual ~50-line skeleton will be moved last.)

Batch 3 partial — Subsystem init extraction:
  - master_node/storage.rs: Phase 1 (db open + stores), returns
    StorageHandles. ~84 lines extracted.
  - master_node/keys.rs: Phase 2 (key management), returns
    KeyHandles. ~42 lines extracted.

main.rs: 6,208 → 4,737 lines (~24% reduction so far). All extracted
modules compile and pass cargo nextest. simtest/ci.sh -short passes.
